Refactor do botão de retornar para pagina anterior
