+++ Frontend/src/pages/Book.jsx
@@ -1,13 +1,14 @@
 import React, { useEffect, useState } from 'react';
-import { Link, useParams, useNavigate } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { api } from '../lib/axios';
 
-import { Container, BackHome, ContainerBook, CardInfo, ContainerButtons, Detail, ContainerDetails } from '../styles/Book';
+import { Container, ContainerBook, CardInfo, ContainerButtons, Detail, ContainerDetails } from '../styles/Book';
 import { Main, H2, Label, Button, Paragraph, Span } from '../styles/styles';
+
+import { BackButton } from '../components/BackButton';
 
 export const Book = () => {
   const { ISBN } = useParams()
-  const navigate = useNavigate();
   const [bookExists, setBookExists] = useState(false)
   const [book, setBook] = useState({})
   
@@ -30,16 +31,9 @@ export const Book = () => {
     <Main>
       {
         bookExists ?  
-
+        
         <Container>
-          <Link onClick={() => navigate(-1)}>
-            <BackHome>
-              <svg width="13" height="21" viewBox="0 0 13 21" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <path d="M0.44264 9.38472L8.99661 0.462512C9.58784 -0.154171 10.5439 -0.154171 11.1288 0.462512L12.5503 1.94517C13.1415 2.56186 13.1415 3.55904 12.5503 4.16917L6.49332 10.5L12.5566 16.8243C13.1478 17.441 13.1478 18.4381 12.5566 19.0483L11.1351 20.5375C10.5439 21.1542 9.58784 21.1542 9.0029 20.5375L0.448929 11.6153C-0.148591 10.9986 -0.148591 10.0014 0.44264 9.38472Z" fill="#619885"/>
-              </svg>
-              <Span fontSize="xxl" fontWeight="600" color="green500">Back</Span>
-            </BackHome>
-          </Link>
+          <BackButton />
 
           <H2 fontSize="xxxl" fontWeight="600">
             Genres : { book.genres.reduce((genreBefore, actualGenre)=>{
