+++ Frontend/src/styles/BackButton.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const BackButton = styled.div`
+    display: flex;
+    align-items: center;
+`;
